+++ src/main/java/uk/gov/hmcts/divorce/caseworker/event/CaseworkerUploadDocument.java
@@ -8,6 +8,7 @@ import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.divorce.common.ccd.PageBuilder;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseDocuments;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
 
@@ -37,7 +38,9 @@ public class CaseworkerUploadDocument implements CCDConfig<CaseData, State, User
             .grantHistoryOnly(SOLICITOR, SUPER_USER, LEGAL_ADVISOR))
             .page("uploadDocument")
             .pageLabel("Upload document")
-            .optional(CaseData::getDocumentsUploaded);
+            .complex(CaseData::getDocuments)
+                .optional(CaseDocuments::getDocumentsUploaded)
+                .done();
     }
 
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(
@@ -48,7 +51,7 @@ public class CaseworkerUploadDocument implements CCDConfig<CaseData, State, User
 
         var caseData = details.getData();
 
-        caseData.sortUploadedDocuments(beforeDetails.getData().getDocumentsUploaded());
+        caseData.getDocuments().sortUploadedDocuments(beforeDetails.getData().getDocuments().getDocumentsUploaded());
 
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
